feat: set up _testing/ and test_helpers/, migrate test infrastructure (#52)
- Unit test stubs (echoCtx) move to src/convex/_testing/ as internalQuery
- E2E fixture mutations (createTenant, createMembership) move to
  src/convex/test_helpers/ as public mutations with IS_TEST guard
- All .test.ts files move from src/convex/ to tests/convex/ with $convex/
  alias imports
- test.setup.ts stays at src/convex/ root
- vitest.config.ts include covers src/**/tests/**/scripts/**
- functions.test.ts refactored: single echoCtx, tests assert on returned
  data instead of testing wrapper behavior through separate stubs
