+++ src/convex/_generated/api.d.ts
@@ -8,15 +8,15 @@
  * @module
  */
 
+import type * as _testing_functions from "../_testing/functions.js";
 import type * as admin from "../admin.js";
 import type * as auth from "../auth.js";
 import type * as functions from "../functions.js";
-import type * as functions_test_helper from "../functions_test_helper.js";
 import type * as http from "../http.js";
 import type * as init from "../init.js";
-import type * as membership_test_helper from "../membership_test_helper.js";
 import type * as memberships from "../memberships.js";
-import type * as tenant_test_helper from "../tenant_test_helper.js";
+import type * as test_helpers_membership from "../test_helpers/membership.js";
+import type * as test_helpers_tenant from "../test_helpers/tenant.js";
 
 import type {
   ApiFromModules,
@@ -25,15 +25,15 @@ import type {
 } from "convex/server";
 
 declare const fullApi: ApiFromModules<{
+  "_testing/functions": typeof _testing_functions;
   admin: typeof admin;
   auth: typeof auth;
   functions: typeof functions;
-  functions_test_helper: typeof functions_test_helper;
   http: typeof http;
   init: typeof init;
-  membership_test_helper: typeof membership_test_helper;
   memberships: typeof memberships;
-  tenant_test_helper: typeof tenant_test_helper;
+  "test_helpers/membership": typeof test_helpers_membership;
+  "test_helpers/tenant": typeof test_helpers_tenant;
 }>;
 
 /**
