--- src/convex/functions_test_helper.ts
@@ -1,15 +1,0 @@
-import { adminQuery, azQuery } from './functions';
-
-export const echoAuthedCtx = azQuery({
-  args: {},
-  handler: async (ctx) => {
-    return { user: ctx.user, tenant: ctx.tenant };
-  },
-});
-
-export const echoAdminCtx = adminQuery({
-  args: {},
-  handler: async (ctx) => {
-    return { user: ctx.user, tenant: ctx.tenant };
-  },
-});
