+++ tests/support/convex.ts
@@ -61,7 +61,7 @@ export async function createTenant(tenant: {
   type: TenantType;
 }): Promise<TestTenant> {
   const client = convexClient();
-  const id = await client.mutation(api.tenant_test_helper.createTenant, tenant);
+  const id = await client.mutation(api.test_helpers.tenant.createTenant, tenant);
   return { ...tenant, _id: id };
 }
 
@@ -71,5 +71,5 @@ export async function createMembership(opts: {
   role: MembershipRole;
 }): Promise<Id<'memberships'>> {
   const client = convexClient();
-  return await client.mutation(api.membership_test_helper.createMembership, opts);
+  return await client.mutation(api.test_helpers.membership.createMembership, opts);
 }
